+++ source/classes/miscAccelerations/miscAccelerations.cpp
@@ -17,6 +17,7 @@
 #include "classes/miscAccelerations/miscAccelerationsRelativisticEffect.h"
 #include "classes/miscAccelerations/miscAccelerationsSolarRadiationPressure.h"
 #include "classes/miscAccelerations/miscAccelerationsAtmosphericDrag.h"
+#include "classes/miscAccelerations/miscAccelerationsAtmosphericDragFromDensityFile.h"
 #include "classes/miscAccelerations/miscAccelerationsAlbedo.h"
 #include "classes/miscAccelerations/miscAccelerationsAntennaThrust.h"
 #include "classes/miscAccelerations/miscAccelerationsFromParametrization.h"
@@ -29,6 +30,7 @@ GROOPS_REGISTER_CLASS(MiscAccelerations, "miscAccelerationsType",
                       MiscAccelerationsSolarRadiationPressure,
                       MiscAccelerationsAlbedo,
                       MiscAccelerationsAtmosphericDrag,
+                      MiscAccelerationsAtmosphericDragFromDensityFile,
                       MiscAccelerationsAntennaThrust,
                       MiscAccelerationsFromParametrization)
 
@@ -43,17 +45,19 @@ MiscAccelerations::MiscAccelerations(Config &config, const std::string &name)
     std::string type;
     while(readConfigChoice(config, name, type, Config::OPTIONAL, "", "non conservative forces acting on satellites"))
     {
-      if(readConfigChoiceElement(config, "relativisticEffect", type, "Relativistic effect (IERS2010)"))
+      if(readConfigChoiceElement(config, "relativisticEffect",             type, "Relativistic effect (IERS2010)"))
        acc.push_back(new MiscAccelerationsRelativisticEffect(config));
-      if(readConfigChoiceElement(config, "solarRadiationPressure", type, "Solar radiation pressure model"))
+      if(readConfigChoiceElement(config, "solarRadiationPressure",         type, "Solar radiation pressure model"))
        acc.push_back(new MiscAccelerationsSolarRadiationPressure(config));
-      if(readConfigChoiceElement(config, "albedo", type, "Albedo radiation"))
+      if(readConfigChoiceElement(config, "albedo",                         type, "Albedo radiation"))
        acc.push_back(new MiscAccelerationsAlbedo(config));
-      if(readConfigChoiceElement(config, "atmosphericDrag", type, "atmospheric drag"))
+      if(readConfigChoiceElement(config, "atmosphericDrag",                type, "atmospheric drag"))
        acc.push_back(new MiscAccelerationsAtmosphericDrag(config));
-      if(readConfigChoiceElement(config, "antennaThrust", type, "antenna thrust"))
+      if(readConfigChoiceElement(config, "atmosphericDragFromDensityFile", type, "atmospheric drag (from density along orbit)"))
+       acc.push_back(new MiscAccelerationsAtmosphericDragFromDensityFile(config));
+      if(readConfigChoiceElement(config, "antennaThrust",                  type, "antenna thrust"))
        acc.push_back(new MiscAccelerationsAntennaThrust(config));
-      if(readConfigChoiceElement(config, "fromParametrization", type, "from a solution vector with given parametrization"))
+      if(readConfigChoiceElement(config, "fromParametrization",            type, "from a solution vector with given parametrization"))
        acc.push_back(new MiscAccelerationsFromParametrization(config));
       endChoice(config);
       if(isCreateSchema(config))